Make sure we generate nice docs for transports as well.
