+++ core-client/Cargo.toml
@@ -19,31 +19,12 @@ categories = [
 ]
 
 [features]
-tls = ["hyper-tls", "http"]
-http = ["hyper"]
-ws = [
-	"websocket",
-	"tokio",
-]
+tls = ["jsonrpc-client-transports/tls"]
+http = ["jsonrpc-client-transports/http"]
+ws = ["jsonrpc-client-transports/ws"]
 
 [dependencies]
-failure = "0.1"
-futures = "0.1.26"
-hyper = { version = "0.12", optional = true }
-hyper-tls = { version = "0.3.2", optional = true }
-jsonrpc-core = { version = "12.0", path = "../core" }
-log = "0.4"
-serde = "1.0"
-serde_json = "1.0"
-tokio = { version = "0.1", optional = true }
-websocket = { version = "0.22", optional = true }
-
-[dev-dependencies]
-assert_matches = "1.1"
-jsonrpc-http-server = { version = "12.0", path = "../http" }
-lazy_static = "1.0"
-env_logger = "0.6"
-tokio = "0.1"
+jsonrpc-client-transports = { version = "12.0", path = "./transports", default-features = false }
 
 [badges]
 travis-ci = { repository = "paritytech/jsonrpc", branch = "master"}
